refactor(training): fold stacks_v2 into neutral stacks module, drop legacy loader
